cleanup: move classes to separate files
